test: 이벤트 테스트 파일 분리
